rename Components folder
